@@ -13,55 +13,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.8",
-    "@angular/common": "^7.2.8",
-    "@angular/compiler": "^7.2.8",
-    "@angular/core": "^7.2.8",
-    "@angular/forms": "^7.2.8",
-    "@angular/http": "^7.2.8",
-    "@angular/platform-browser": "^7.2.8",
-    "@angular/platform-browser-dynamic": "^7.2.8",
-    "@angular/router": "^7.2.8",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.15",
-    "@fortawesome/free-solid-svg-icons": "^5.7.2",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.0",
-    "@ng-select/ng-select": "^2.20.0",
+    "@angular-devkit/build-ng-packagr": "~0.801.2",
+    "@angular/animations": "^8.1.2",
+    "@angular/common": "^8.1.2",
+    "@angular/compiler": "^8.1.2",
+    "@angular/core": "^8.1.2",
+    "@angular/forms": "^8.1.2",
+    "@angular/platform-browser": "^8.1.2",
+    "@angular/platform-browser-dynamic": "^8.1.2",
+    "@angular/router": "^8.1.2",
+    "@fortawesome/angular-fontawesome": "^0.4.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.19",
+    "@fortawesome/free-solid-svg-icons": "^5.9.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.0",
+    "@ng-select/ng-select": "^3.0.3",
     "@types/hashids": "^1.0.30",
     "bootstrap": "^4.3.1",
-    "camera-controls": "^1.12.0",
-    "core-js": "^2.6.5",
+    "camera-controls": "^1.12.2",
+    "core-js": "^3.1.4",
     "drag-controls": "^1.0.4",
     "file-saver": "^2.0.2",
     "hashids": "^1.2.2",
-    "ng5-slider": "^1.1.14",
-    "rxjs": "^6.4.0",
-    "three": "^0.102.1",
-    "zone.js": "^0.8.29"
+    "ng5-slider": "^1.2.4",
+    "rxjs": "^6.5.2",
+    "three": "^0.106.2",
+    "tsickle": "^0.36.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.5",
-    "@angular/compiler-cli": "^7.2.8",
-    "@angular/language-service": "^7.2.8",
-    "@types/jasmine": "~3.3.9",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular/cli": "^8.1.2",
+    "@angular/compiler-cli": "^8.1.2",
+    "@angular/language-service": "^8.1.2",
+    "@types/jasmine": "~3.3.15",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~11.11.3",
-    "@types/three": "0.93.26",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~3.3.0",
+    "@types/node": "~12.6.8",
+    "@types/three": "^0.103.2",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^4.7.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^5.1.0",
     "protractor": "~5.4.2",
-    "ts-node": "~8.0.3",
-    "tsickle": "^0.34.3",
-    "tslib": "^1.9.3",
-    "tslint": "~5.14.0",
-    "typescript": "3.2.4"
+    "ts-node": "~8.3.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.18.0",
+    "typescript": "^3.4.5"
   }
 }
